adds user-dialog-modals accross entire application

+++ frontend/src/app/daycare-group/daycare-group-page/daycare-group-page.component.ts
@@ -2,9 +2,11 @@ import { Component, OnInit } from '@angular/core';
 import { DaycareGroup } from '../daycare-group';
 import { DaycareGroupDataService } from '../daycare-group-data.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
+import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal';
 import { DaycareGroupCreateEditComponent } from '../daycare-group-create-edit/daycare-group-create-edit.component';
 import { DialogModalService } from 'src/app/dialog/dialog-modal.service';
+import { Child } from 'src/app/child/child';
+import { CONFIRMATION_HEADER, ERROR_HEADER } from 'src/app/const';
 
 @Component({
   selector: 'app-daycare-group-page',
@@ -13,7 +15,7 @@ import { DialogModalService } from 'src/app/dialog/dialog-modal.service';
 })
 export class DaycareGroupPageComponent implements OnInit {
 
-  private bsModalRef: BsModalRef
+  private modalRef: BsModalRef
 
   private daycareGroup: DaycareGroup
 
@@ -49,14 +51,14 @@ export class DaycareGroupPageComponent implements OnInit {
 
   openEditGroupModal() {
     let initialState = { daycareGroup: new DaycareGroup(this.daycareGroup.id, this.daycareGroup.groupName) }
-    this.bsModalRef = this.bsModalService.show(DaycareGroupCreateEditComponent,
+    this.modalRef = this.bsModalService.show(DaycareGroupCreateEditComponent,
       { class: 'modal-top-10 modal-sm', initialState, ignoreBackdropClick: true })
-    this.bsModalRef.content.onClose.subscribe(
+    this.modalRef.content.onClose.subscribe(
       onClose => {
         if (onClose) {
-          this.bsModalRef = this.dialogModalService.openInformationModal('Group edited',
+          this.modalRef = this.dialogModalService.openInformationModal('Group edited',
             `Daycare group#${this.daycareGroup.id} has been succesfully edited.`)
-          this.bsModalRef.content.onClose.subscribe(
+          this.modalRef.content.onClose.subscribe(
             onClose => {
               this.retrieveDaycareGroup(this.daycareGroup.id)
             })
@@ -65,26 +67,47 @@ export class DaycareGroupPageComponent implements OnInit {
   }
 
   deleteGroup() {
-    this.daycareGroupDataService.deleteDaycareGroup(this.daycareGroup.id).subscribe(
-      response => {
-        this.bsModalRef = this.dialogModalService.openInformationModal('Group deleted',
-          `Daycare group#${this.daycareGroup.id} has been succesfully deleted.`)
-        this.bsModalRef.content.onClose.subscribe(
-          onClose => {
-            this.router.navigate(['daycare-group-list'])
-          })
+    this.modalRef = this.dialogModalService.openConfirmationModal(CONFIRMATION_HEADER,
+      `You are about to delete daycare group#${this.daycareGroup.id}.\n
+      Any children, that are currently assigned to it will be left with no group.`)
+    this.modalRef.content.onClose.subscribe(
+      onClose => {
+        if (onClose) {
+          this.daycareGroupDataService.deleteDaycareGroup(this.daycareGroup.id).subscribe(
+            response => {
+              this.modalRef = this.dialogModalService.openInformationModal('Group deleted',
+                `Daycare group#${this.daycareGroup.id} has been succesfully deleted.`)
+              this.modalRef.content.onClose.subscribe(
+                onClose => {
+                  this.router.navigate(['daycare-group-list'])
+                })
+            },
+            err => {
+              this.dialogModalService.openInformationModal(ERROR_HEADER, err.message)
+            })
+        }
       })
   }
 
-  removeChildFromGroup(childId: number) {
-    this.daycareGroupDataService.removeChildFromDaycareGroup(this.daycareGroup.id, childId).subscribe(
-      response => {
-        this.bsModalRef = this.dialogModalService.openInformationModal('Child removed',
-          `Child#${childId} has been succesfully removed from daycare group#${this.daycareGroup.id}.`)
-        this.bsModalRef.content.onClose.subscribe(
-          onclose => {
-            this.retrieveDaycareGroup(this.daycareGroup.id)
-          })
+  removeChildFromGroup(child: Child) {
+    this.modalRef = this.dialogModalService.openConfirmationModal(CONFIRMATION_HEADER,
+      `You are about to remove child#${child.id} (${child.firstName} ${child.lastName}) from daycare group#${this.daycareGroup.id}`)
+    this.modalRef.content.onClose.subscribe(
+      onclose => {
+        if (onclose) {
+          this.daycareGroupDataService.removeChildFromDaycareGroup(this.daycareGroup.id, child.id).subscribe(
+            response => {
+              this.modalRef = this.dialogModalService.openInformationModal('Child removed',
+                `Child#${child.id} (${child.firstName} ${child.lastName}) has been succesfully removed from daycare group#${this.daycareGroup.id}.`)
+              this.modalRef.content.onClose.subscribe(
+                onclose => {
+                  this.retrieveDaycareGroup(this.daycareGroup.id)
+                })
+            },
+            err => {
+              this.dialogModalService.openInformationModal(ERROR_HEADER, err.message)
+            })
+        }
       })
   }
